pyproject.toml

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imdb-recommender"
version = "0.1.0"
description = "Local IMDb recommender with rating/watchlist logging for Selenium replay"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="IMDb Recommender Authors"}]
dependencies = [
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "pydantic>=2.4.0",
  "typer>=0.9.0",
  "click>=8.1.3",
  "python-dotenv>=1.0.0",
  "tomli>=2.0.1; python_version<'3.11'",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "hypothesis>=6.82.0",
  "selenium>=4.15.0",
  "webdriver-manager>=4.0.0",
]
dev = [
  "ruff>=0.1.6",
  "black>=23.9.1", 
  "pre-commit>=3.5.0",
]
all = [
  "pyarrow>=13.0.0",
  "scikit-learn>=1.3.0",
  "openpyxl>=3.1.0",
]

[project.scripts]
imdbrec = "imdb_recommender.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/imdb_recommender"]

[tool.hatch.build.targets.sdist]
include = ["src/imdb_recommender"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["src"]
